format code create_db, queries

+++ src/com/startjava/lesson_5/create_db.sql
@@ -3,12 +3,12 @@ CREATE DATABASE jaegers;
 \connect jaegers;
 
 CREATE TABLE jaegers (
-    id SERIAL PRIMARY KEY, 
-    modelName TEXT, 
-    mark TEXT, 
-    height INTEGER, 
-    weight DOUBLE PRECISION, 
-    status BOOLEAN, 
-    origin TEXT, 
-    launch INTEGER, 
-    kaijuKill INTEGER);
+    id          SERIAL PRIMARY KEY, 
+    modelName   TEXT, 
+    mark        TEXT, 
+    height      INTEGER, 
+    weight      REAL, 
+    status      BOOLEAN, 
+    origin      TEXT, 
+    launch      INTEGER, 
+    kaijuKill   INTEGER);
